Move get gpu function to core
